+++ org/postgresql/jdbc2/AbstractJdbc2ResultSet.java
@@ -3,7 +3,7 @@
 * Copyright (c) 2003-2005, PostgreSQL Global Development Group
 *
 * IDENTIFICATION
-*   $PostgreSQL: pgjdbc/org/postgresql/jdbc2/AbstractJdbc2ResultSet.java,v 1.80.2.1 2005/12/04 21:41:21 jurka Exp $
+*   $PostgreSQL: pgjdbc/org/postgresql/jdbc2/AbstractJdbc2ResultSet.java,v 1.80.2.2 2006/08/06 18:11:43 jurka Exp $
 *
 *-------------------------------------------------------------------------
 */
@@ -2062,7 +2062,7 @@ public abstract class AbstractJdbc2ResultSet implements BaseResultSet, org.postg
                 if ( fields[columnIndex - 1].getOID() == 26)
                 {
                     LargeObjectManager lom = connection.getLargeObjectAPI();
-                    LargeObject lob = lom.open(getInt(columnIndex));
+                    LargeObject lob = lom.open(getLong(columnIndex));
                     byte buf[] = lob.read(lob.size());
                     lob.close();
                     return trimBytes(columnIndex, buf);
@@ -2178,7 +2178,7 @@ public abstract class AbstractJdbc2ResultSet implements BaseResultSet, org.postg
             if ( fields[columnIndex - 1].getOID() == 26)
             {
                 LargeObjectManager lom = connection.getLargeObjectAPI();
-                LargeObject lob = lom.open(getInt(columnIndex));
+                LargeObject lob = lom.open(getLong(columnIndex));
                 return lob.getInputStream();
             }
         }
